+++ src/StefanoTree/NestedSet/Adapter/Doctrine2DBAL.php
@@ -539,7 +539,7 @@ class Doctrine2DBAL implements AdapterInterface
         return (is_array($result)) ? $result : array();
     }
 
-    public function getDescendants($nodeId = 1, $startLevel = 0, $levels = null, $excludeBranch = null)
+    public function getDescendants($nodeId, $startLevel = 0, $levels = null, $excludeBranch = null)
     {
         $options = $this->getOptions();