+++ src/StefanoTree/NestedSet/Adapter/Zend1.php
@@ -473,7 +473,7 @@ class Zend1 implements AdapterInterface
         return $result;
     }
 
-    public function getDescendants($nodeId = 1, $startLevel = 0, $levels = null, $excludeBranch = null)
+    public function getDescendants($nodeId, $startLevel = 0, $levels = null, $excludeBranch = null)
     {
         $options = $this->getOptions();